requirements.txt
streamlit
openai
python-docx
PyMuPDF
groq
python-dotenv